backend/requirements.txt
fastapi
uvicorn
pydantic-settings
langchain
langchain-openai
langchain-community
dashscope
chromadb
python-dotenv
openai
requests
python-multipart
